grid.pattern_magick {gridpattern}	R Documentation

Magick patterned grobs

Description

grid.pattern_magick() draws a imagemagick pattern onto the graphic device. names_magick, names_magick_intensity, and names_magick_stripe are character vectors of supported type values plus subsets for shaded intensity and stripes.

Usage

grid.pattern_magick(
  x = c(0, 0, 1, 1),
  y = c(1, 0, 0, 1),
  id = 1L,
  ...,
  type = "hexagons",
  fill = "grey20",
  scale = 1,
  filter = "box",
  alpha = gp$alpha %||% NA_real_,
  aspect_ratio = 1,
  key_scale_factor = 1,
  res = getOption("ggpattern_res", 72),
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)

names_magick

names_magick_intensity

names_magick_stripe

Arguments

x
A numeric vector or unit object specifying x-locations of the pattern boundary.

y
A numeric vector or unit object specifying y-locations of the pattern boundary.

id
A numeric vector used to separate locations in x, y into multiple boundaries. All locations within the same id belong to the same boundary.

...
Currently ignored.

type
Magick pattern types. names_magick, names_magick_intensity, and names_magick_stripe are character vectors of supported type values plus subsets for shaded intensity and stripes.

fill
Fill colour

scale
Extra scaling

filter
Filter to use when scaling. magick::filter_types() returns a vector of supported values.

alpha
Alpha (between 0 and 1) or NA (default, preserves colors' alpha value).

aspect_ratio
Override aspect ratio.

key_scale_factor
Additional scale factor for legend.

res
Assumed resolution (in pixels per graphic device inch) to use when creating array pattern.

default.units
A string indicating the default units to use if x or y are only given as numeric vectors.

name
A character identifier.

gp
An object of class "gpar", typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

draw
A logical value indicating whether graphics output should be produced.

vp
A Grid viewport object (or NULL).

Format

An object of class character of length 54.

An object of class character of length 21.

An object of class character of length 19.

Value

A grid grob object invisibly. If draw is TRUE then also draws to the graphic device as a side effect.

See Also

The imagemagick documentation http://www.imagemagick.org/script/formats.php for more information.

Examples

  if (requireNamespace("magick")) {
    x_hex <- 0.5 + 0.5 * cos(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
    y_hex <- 0.5 + 0.5 * sin(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
    grid.pattern_magick(x_hex, y_hex, type="octagons", fill="blue", scale=2)
  }

  # supported magick pattern names
  print(names_magick)

[Package gridpattern version 1.3.1 Index]