mean_col {gridpattern}	R Documentation

Compute average color

Description

mean_col() computes an average color.

Usage

mean_col(...)

Arguments

...
Colors to average

Details

We currently compute an average color by using the quadratic mean of the colors' RGBA values.

Value

A color string of 9 characters: "#" followed by the red, blue, green, and alpha values in hexadecimal.

Examples

 mean_col("black", "white")
 mean_col(c("black", "white"))
 mean_col("red", "blue")

[Package gridpattern version 1.3.1 Index]